Adding a check for the address.address in on listen event
Motivation:
When trying out the example given in the README.MD the following is
displayed upon start up:
Cluster worker 1 is now listening at null:8080

The address.address is null when the master receives the listen event.
The port and type of the address are given though.

Modifications:
As a suggestion the check could be added. Another option is to add the
host address specifically to the app.listen(port, hostname) to the
example snippet.

Result:
The logs will now be:

Cluster worker 1 is now listening at 0.0.0.0:8080
One line description of your change