Spawn new worker on disconnect instead of exit
There are some situations where worker.disconnect might be called, but
not worker.kill. The disconnect event will be emitted for both, so it's
better if we re-spawn based on this.

I've also added a condition that the cluster master should do this,
since the disconnect event will get emitted for both the master and
workers.

This change also puts the logging of the backoff message in the correct
place -- the message should appear before the backoff period, rather
than at the end of it.